final commits

+++ src/main/java/com/tshepo/SPRING/BOOT/PART1/service/UserService.java
@@ -1,7 +1,7 @@
 package com.tshepo.SPRING.BOOT.PART1.service;
 
 public interface UserService {
-    String addUser(int userId, String userName, String userSurname) throws NoSuchMethodException;
-    String getUser(int userId);
-    String removeUser(int userId);
+    String addUser(long userId, String userName, String userSurname) throws NoSuchMethodException;
+    String getUser(long userId);
+    String removeUser(long userId);
 }
